stalling issue
currently the code sometimes stalls in cli_utils.evaluate_model during "Computing metrics"
the traceback goes to dataloader even though it is not used in the computing metrics at alll
I decided to delete the lightning_module (as it is loaded every time anyways) so that there wouldn't be any references to the dataloader.

Also, to have a better debugging options, I expanded the logging for metrics computation. Now every evaluated example is logged into "debug.log" file.